+++ jeap-spring-boot-db-migration-starter-it/src/test/java/ch/admin/bit/jeap/db/migration/FlywayMigrationConfigurationIT.java
@@ -11,7 +11,6 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.context.bean.override.mockito.MockReset;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -21,7 +20,6 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 
 @SpringBootTest(classes = TestConfig.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Testcontainers
 public class FlywayMigrationConfigurationIT extends PostgresTestContainerBase {
 
     private static final String MIGRATED_SCHEMA_VERSION = "1.0.0";