+++ jeap-spring-boot-db-migration-starter-it/src/test/java/ch/admin/bit/jeap/db/migration/PostgresTestContainerBase.java
@@ -2,15 +2,19 @@ package ch.admin.bit.jeap.db.migration;
 
 import org.springframework.test.context.DynamicPropertyRegistry;
 import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.postgresql.PostgreSQLContainer;
+import org.testcontainers.utility.DockerImageName;
 
-public class PostgresTestContainerBase {
+@Testcontainers
+public abstract class PostgresTestContainerBase {
 
-    private static PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:17-alpine");
-
-    static {
-        postgres.start();
-    }
+    @Container
+    static final PostgreSQLContainer postgres = new PostgreSQLContainer(DockerImageName.parse("postgres:17-alpine"))
+            .withDatabaseName("test")
+            .withUsername("test")
+            .withPassword("test");
 
     @DynamicPropertySource
     static void configureProperties(DynamicPropertyRegistry registry) {